// src/sdk/generated/models/BulkInstitutionOnboardResult.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type BulkInstitutionOnboardResult = {
  total: number;
  success: number;
  failed: number;
  errors: Array<Record<string, any>>;
  created_institution_ids: Array<string>;
};